Get slim bongo cat to work.
Total weight for this feature: 7332 bytes.
That is about a 30% improvement over the original bongo_cat in j-inc's
keymap.

It took some doing, including a modification to the oled driver.

Code isn't clean yet. I just got it to work.
It was harder than I thought with the subtleties of working with PROGMEM
and SRAM.

The more compact bytes are generated from a rust program that I will
clean-up and share in due time.